Refactoring: 함께하기 우선 수정 (#53)

+++ src/main/java/com/back/domain/together/dto/request/TogetherRequest.java
@@ -32,7 +32,7 @@ public class TogetherRequest {
     private TogetherMode mode;
 
     @NotNull
-    private Integer capacity;
+    private Long capacity;
 
     @NotNull
     private LocalDate startDate;
